refactor(core): split out test schema
